requirements.txt
Click>=3.1
six>=1.7.3
Flask-Mako>=0.3
Flask>=0.10.1
Frozen-Flask>=0.11
Mako>=1.0.0
PyYAML>=3.11
dulwich>=0.9.0
feedparser>=5.1.3
sh>=1.09
tornado>=4.0.1
